formats/polus-vector-converter-plugins/src/requirements.txt
filepattern==1.4.7
bfio[all]==2.1.9
scipy==1.7.1
cupy-cuda114[all]==9.6.0
preadator==0.2.0
